Clean up default config behaviour
