+++ configs/mappings/shared_mappings.yaml
@@ -4,11 +4,12 @@ code_systems:
     "2.16.840.1.113883.6.96": "http://snomed.info/sct"
     "2.16.840.1.113883.6.1": "http://loinc.org"
     "2.16.840.1.113883.6.88": "http://www.nlm.nih.gov/research/umls/rxnorm"
+    "2.16.840.1.113883.3.26.1.1": "http://ncicb.nci.nih.gov/xml/owl/EVS/Thesaurus.owl"
   fhir_to_cda:
     "http://snomed.info/sct": "2.16.840.1.113883.6.96"
     "http://loinc.org": "2.16.840.1.113883.6.1"
     "http://www.nlm.nih.gov/research/umls/rxnorm": "2.16.840.1.113883.6.88"
-    "http://terminology.hl7.org/CodeSystem/condition-clinical": "2.16.840.1.113883.6.96"
+    "http://ncicb.nci.nih.gov/xml/owl/EVS/Thesaurus.owl": "2.16.840.1.113883.3.26.1.1"
 
 status_codes:
   cda_to_fhir:
