+++ configs/templates/cda_fhir/cda_medication_entry.liquid
@@ -7,7 +7,9 @@
       {"@root": "{{template_id}}"} {% if forloop.last != true %},{% endif %}
     {% endfor %}
     ],
-    "id": {"@root": "{{ resource.id | generate_id }}"},
+    {% if resource.id %}
+    "id": {"@root": "{{ resource.id }}"},
+    {% endif %}
     "statusCode": {"@code": "{{ config.template.substance_admin.status_code }}"},
     {% if resource.dosage and resource.dosage[0].doseAndRate %}
     "doseQuantity": {
