+++ configs/templates/cda_fhir/condition.liquid
@@ -1,25 +1,25 @@
 {
   "resourceType": "Condition",
   {% if entry.act.entryRelationship.is_array %}
-    {% assign actEntryRelationship = entry.act.entryRelationship[0] %}
+    {% assign obs = entry.act.entryRelationship[0].observation %}
   {% else %}
-    {% assign actEntryRelationship = entry.act.entryRelationship %}
+    {% assign obs = entry.act.entryRelationship.observation %}
   {% endif %}
-  {% if actEntryRelationship.observation.entryRelationship.observation.code['@code'] == config.identifiers.clinical_status.code %}
-    {% if actEntryRelationship.observation.entryRelationship.observation.value %}
+  {% if obs.entryRelationship.observation.code['@code'] == config.identifiers.clinical_status.code %}
+    {% if obs.entryRelationship.observation.value %}
     "clinicalStatus": {
       "coding": [
         {
           "system": "http://terminology.hl7.org/CodeSystem/condition-clinical",
-          "code": "{{ actEntryRelationship.observation.entryRelationship.observation.value['@code'] | map_status: 'cda_to_fhir' }}"
+          "code": "{{ obs.entryRelationship.observation.value['@code'] | map_status: 'cda_to_fhir' }}"
         },
         {
-          "system": "{{ actEntryRelationship.observation.entryRelationship.observation.value['@codeSystem'] | map_system: 'cda_to_fhir' }}",
-          "code": "{{ actEntryRelationship.observation.entryRelationship.observation.value['@code'] }}",
-          "display": "{{ actEntryRelationship.observation.entryRelationship.observation.value['@displayName'] }}"
+          "system": "{{ obs.entryRelationship.observation.value['@codeSystem'] | map_system: 'cda_to_fhir' }}",
+          "code": "{{ obs.entryRelationship.observation.value['@code'] }}",
+          "display": "{{ obs.entryRelationship.observation.value['@displayName'] }}"
         }
       ]
-    },
+    }{% if true %},{% endif %}
     {% endif %}
   {% endif %}
   "category": [{
@@ -28,25 +28,22 @@
       "code": "problem-list-item",
       "display": "Problem List Item"
     }]
-  }],
-  {% if actEntryRelationship.observation.value %}
+  }]{% if obs.value or obs.effectiveTime %},{% endif %}
+  {% if obs.value %}
   "code": {
     "coding": [{
-      "system": "{{ actEntryRelationship.observation.value['@codeSystem'] | map_system: 'cda_to_fhir' }}",
-      "code": "{{ actEntryRelationship.observation.value['@code'] }}",
-      "display": "{{ actEntryRelationship.observation.value['@displayName'] }}"
+      "system": "{{ obs.value['@codeSystem'] | map_system: 'cda_to_fhir' }}",
+      "code": "{{ obs.value['@code'] }}",
+      "display": "{{ obs.value['@displayName'] }}"
     }]
-  },
+  }{% if obs.effectiveTime %},{% endif %}
   {% endif %}
-  {% if actEntryRelationship.observation.effectiveTime %}
-    {% if actEntryRelationship.observation.effectiveTime.low %}
-    "onsetDateTime": "{{ actEntryRelationship.observation.effectiveTime.low['@value'] | format_date }}",
+  {% if obs.effectiveTime %}
+    {% if obs.effectiveTime.low %}
+    "onsetDateTime": "{{ obs.effectiveTime.low['@value'] | format_date }}"{% if obs.effectiveTime.high %},{% endif %}
     {% endif %}
-    {% if actEntryRelationship.observation.effectiveTime.high %}
-    "abatementDateTime": "{{ actEntryRelationship.observation.effectiveTime.high['@value'] | format_date }}",
+    {% if obs.effectiveTime.high %}
+    "abatementDateTime": "{{ obs.effectiveTime.high['@value'] | format_date }}"
     {% endif %}
   {% endif %}
-  "subject": {
-    "reference": "Patient/{{ entry.subject_id | default: 'Foo' }}"
-  }
 }
